requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.3
httpx==0.27.2
email-validator==2.1.0
bcrypt==4.0.1
cryptography
stripe==14.3.0
apscheduler==3.10.4
xhtml2pdf
chromadb==0.5.23
sentence-transformers==3.4.1
aiosmtplib==3.0.2
gunicorn==21.2.0